Cache page props and fix a problem with precaching
Precaching was used to make the web app work offline, but it precached everyting,
which led to huge cache size.
Configured the workbox webpack plugin to cache images at runtime only
and to only cache the first http request.

Aside from working offline a simple cache strategy is implemented,
to make navigating to the same page withing a short period of time fast.