Add VirtualList component to improve performance
Add a new VirtualList component that skips rendering off-screen items.
The component replaces off-screen items with two divs.
One for items before the first index and one for those after that last.
Their mission is to take the height of the items they represent.
Refs are used in place of state for everything.
This is to take full control of when exactly to rerender.
This allows for some optimizations in the implementation.
I opted for this approach instead of absolutely positiones items,
because this approach doesn't have to rerender on every scroll.